Deleted CounterFragment.setIcons(), and made updateCounterIconTab*() local methods.
